Refactor rsvp dialog and use in nav as well
