+++ src/app/components/common/simple-dialog.js
