+++ src/app/containers/cta-section.js
@@ -1,6 +1,5 @@
-import React from 'react';
 import { connect } from 'react-redux';
-import { showModal } from '../actions/show-modal';
+import { showRsvpModal } from '../actions/show-rsvp-modal';
 import Cta from '../components/cta';
 
 const mapStateToProps = () => {
@@ -10,13 +9,7 @@ const mapStateToProps = () => {
 const mapDispatchToProps = (dispatch) => {
   return {
     onRsvpClick: () => {
-      dispatch(showModal({
-        modalType: 'rsvp',
-        modalProps: {
-          title: 'RSVP Now',
-          content: <div class="hello">Hewo</div>,
-        },
-      }));
+      dispatch(showRsvpModal());
     },
   };
 };
